requirements.txt
fastparquet==0.8.1
pyarrow==7.0.0
pandas==1.1.5
numpy==1.21.6
torch==1.11.0
matplotlib==3.5.1
lightgbm==3.3.2
torchvision==0.12.0
Pillow==9.1.0
recommenders==1.1.0
cornac==1.14.2
scrapbook==0.5.0
hyperopt==0.2
pickle5==0.0.12
cython==0.29.28
bson==0.5.10
hyperas==0.4.1
faiss==1.5.3
git+https://github.com/mayukh18/reco
